Reuse shared CLI release resolver for upgrades
- Extract latest-release lookup into a shared effect
- Use channel-aware GitHub release filtering for install and upgrade flows
- Keep binary upgrade error mapping aligned with release resolution failures
