+++ ts/packages/cli/src/effects/install-skill.ts
@@ -6,26 +6,18 @@ import { NodeOs } from 'src/services/node-os';
 import { TerminalUI } from 'src/services/terminal-ui';
 import { GITHUB_CONFIG } from 'src/effects/github-config';
 import { APP_VERSION } from 'src/constants';
-import { semverComparator } from 'src/effects/compare-semver';
+import {
+  fetchLatestCliRelease,
+  type GitHubRelease,
+  type GitHubRepoConfig,
+} from 'src/effects/resolve-cli-release';
 import decompress from 'decompress';
 
 const SKILL_NAME = 'composio-cli';
 const SKILL_ASSET_NAME = 'composio-skill.zip';
-const CLI_RELEASE_TAG_PATTERN = /^@composio\/cli@\d+\.\d+\.\d+.*$/;
-
 export type SkillReleaseChannel = 'beta' | 'stable';
 
-type GitHubRelease = {
-  tag_name: string;
-  prerelease?: boolean;
-  draft?: boolean;
-  assets: Array<{ name: string; browser_download_url: string }>;
-};
-
-type GitHubConfig = {
-  API_BASE_URL: string;
-  OWNER: string;
-  REPO: string;
+type GitHubConfig = GitHubRepoConfig & {
   TAG: Option.Option<string>;
   ACCESS_TOKEN: Option.Option<string>;
 };
@@ -61,61 +53,13 @@ export const resolveSkillReleaseTag = ({
       return `@composio/cli@${APP_VERSION}`;
     }
 
-    const releaseUrl = `${githubConfig.API_BASE_URL}/repos/${githubConfig.OWNER}/${githubConfig.REPO}/releases?per_page=100`;
-    const releaseResponse = yield* httpClient
-      .get(releaseUrl)
-      .pipe(
-        Effect.catchAll(error =>
-          Effect.fail(new Error(`Failed to fetch ${channel} releases from GitHub: ${error}`))
-        )
-      );
-
-    if (releaseResponse.status < 200 || releaseResponse.status >= 300) {
-      return yield* Effect.fail(
-        new Error(
-          `Failed to fetch ${channel} releases from GitHub (HTTP ${releaseResponse.status})`
-        )
-      );
-    }
-
-    const releases = (yield* releaseResponse.json.pipe(
-      Effect.catchAll(() => Effect.fail(new Error('Failed to parse GitHub releases JSON')))
-    )) as unknown;
-
-    if (!Array.isArray(releases)) {
-      return yield* Effect.fail(new Error('GitHub releases response was not an array'));
-    }
-
-    const prerelease = channel === 'beta';
-    const matchingReleases = releases.filter(
-      (release): release is GitHubRelease =>
-        typeof release === 'object' &&
-        release !== null &&
-        'tag_name' in release &&
-        typeof release.tag_name === 'string' &&
-        ('prerelease' in release ? release.prerelease === prerelease : prerelease === false) &&
-        ('draft' in release ? release.draft === false : true) &&
-        CLI_RELEASE_TAG_PATTERN.test(release.tag_name) &&
-        Array.isArray(release.assets) &&
-        hasSkillAsset(release)
-    );
-
-    if (matchingReleases.length === 0) {
-      return yield* Effect.fail(
-        new Error(`No ${channel} CLI releases found with ${SKILL_ASSET_NAME}`)
-      );
-    }
-
-    let latest = matchingReleases[0];
-    for (const release of matchingReleases.slice(1)) {
-      const comparison = yield* semverComparator(latest.tag_name, release.tag_name).pipe(
-        Effect.mapError(error => new Error(`Failed to compare release versions: ${error}`))
-      );
-
-      if (comparison < 0) {
-        latest = release;
-      }
-    }
+    const latest = yield* fetchLatestCliRelease({
+      assetDescription: SKILL_ASSET_NAME,
+      channel,
+      githubConfig,
+      hasRequiredAsset: hasSkillAsset,
+      httpClient,
+    });
 
     return latest.tag_name;
   });
